Break functions and other improvements
