+++ pysd/__init__.py
@@ -1,5 +1,5 @@
 from .pysd import read_vensim, read_xmile, load
-from .py_backend import functions, utils, external
+from .py_backend import functions, statefuls, utils, external
 from .py_backend.decorators import cache, subs
 from ._version import __version__
 from .py_backend.vensim.table2py import read_tabular
