+++ tests/unit_test_vensim2py.py
@@ -380,12 +380,17 @@ class TestParse_general_expression(unittest.TestCase):
     def test_function_calls(self):
         from pysd.py_backend.vensim.vensim2py import parse_general_expression
 
-        res = parse_general_expression({"expr": "ABS(StockA)"},
+        res = parse_general_expression({"expr": "ABS(StockA)",
+                                        "real_name": "AB",
+                                        "eqn": "AB = ABS(StockA)"},
                                        {"StockA": "stocka"})
         self.assertEqual(res[0]["py_expr"], "abs(stocka())")
 
         res = parse_general_expression(
-            {"expr": "If Then Else(A>B, 1, 0)"}, {"A": "a", "B": "b"}
+            {"expr": "If Then Else(A>B, 1, 0)",
+             "real_name": "IFE",
+             "eqn": "IFE = If Then Else(A>B, 1, 0)"},
+            {"A": "a", "B": "b"}
         )
         self.assertEqual(
             res[0]["py_expr"], "if_then_else(a()>b(), lambda: 1, lambda: 0)"
@@ -481,7 +486,7 @@ class TestParse_general_expression(unittest.TestCase):
         """ stock construction should create a stateful variable and
         reference it """
         from pysd.py_backend.vensim.vensim2py import parse_general_expression
-        from pysd.py_backend.functions import Integ
+        from pysd.py_backend.statefuls import Integ
 
         res = parse_general_expression(
             {
@@ -502,7 +507,7 @@ class TestParse_general_expression(unittest.TestCase):
 
     def test_delay_construction_function_no_subscripts(self):
         from pysd.py_backend.vensim.vensim2py import parse_general_expression
-        from pysd.py_backend.functions import Delay
+        from pysd.py_backend.statefuls import Delay
 
         res = parse_general_expression(
             {
@@ -535,7 +540,7 @@ class TestParse_general_expression(unittest.TestCase):
         forecast elements, and then pass back a reference to that value
         """
         from pysd.py_backend.vensim.vensim2py import parse_general_expression
-        from pysd.py_backend.functions import Forecast
+        from pysd.py_backend.statefuls import Forecast
 
         res = parse_general_expression(
             {
@@ -564,7 +569,7 @@ class TestParse_general_expression(unittest.TestCase):
         elements, and then pass back a reference to that value
         """
         from pysd.py_backend.vensim.vensim2py import parse_general_expression
-        from pysd.py_backend.functions import Smooth
+        from pysd.py_backend.statefuls import Smooth
 
         res = parse_general_expression(
             {
@@ -984,6 +989,56 @@ class TestParse_general_expression(unittest.TestCase):
 
         self.assertIn(
             "abs(upper_var())",
+            res[0]["py_expr"], )
+
+    def test_random_0_1(self):
+        from pysd.py_backend.vensim.vensim2py import parse_general_expression
+
+        # When parsing functions arguments first the subscript ranges are
+        # parsed and later the general id is used, however, the if a reference
+        # to a var starts with a subscript range name this could make the
+        # parser crash
+        res = parse_general_expression(
+            {
+                "expr": "RANDOM 0 1()",
+                "real_name": "A",
+                "eqn": "A = RANDOM 0 1()",
+                "py_name": "a",
+                "merge_subs": [],
+                "dependencies": set()
+            },
+            {
+                "A": "a",
+            }
+        )
+
+        self.assertIn(
+            "np.random.uniform(0, 1)",
+            res[0]["py_expr"], )
+
+    def test_random_uniform(self):
+        from pysd.py_backend.vensim.vensim2py import parse_general_expression
+
+        # When parsing functions arguments first the subscript ranges are
+        # parsed and later the general id is used, however, the if a reference
+        # to a var starts with a subscript range name this could make the
+        # parser crash
+        res = parse_general_expression(
+            {
+                "expr": "RANDOM UNIFORM(10, 15, 3)",
+                "real_name": "A",
+                "eqn": "A = RANDOM UNIFORM(10, 15, 3)",
+                "py_name": "a",
+                "merge_subs": [],
+                "dependencies": set()
+            },
+            {
+                "A": "a",
+            }
+        )
+
+        self.assertIn(
+            "np.random.uniform(10, 15)",
             res[0]["py_expr"], )
 
     def test_incomplete_expression(self):
